+++ test/support/web/commanders/shared1_commander.ex
@@ -7,6 +7,7 @@ defmodule DrabTestApp.Shared1Commander do
 
   def button_clicked(socket, sender, arg \\ "outside") do
     # IO.inspect(socket)
+    # IO.inspect this_commander(sender)
     set_prop(socket, this_commander(sender) <> " .spaceholder1", innerText: "changed")
     poke(socket, text: "changed in shared commander, " <> arg, bgcolor: "#77dddd", color: "#990000")
   end
@@ -14,8 +15,4 @@ defmodule DrabTestApp.Shared1Commander do
   defhandler peek_text(socket, sender, _) do
     set_prop(socket, this(sender), innerText: peek(socket, :text) || "--- nil ---")
   end
-  # def page_loaded(socket) do
-  #   DrabTestApp.IntegrationCase.add_page_loaded_indicator(socket)
-  #   DrabTestApp.IntegrationCase.add_pid(socket)
-  # end
 end