+++ test/support/web/commanders/shared2_commander.ex
@@ -2,14 +2,15 @@ defmodule DrabTestApp.Shared2Commander do
   @moduledoc false
 
   use Drab.Commander
-  # onload(:page_loaded)
-  # public :button_clicked
+  public(:button_clicked)
 
-  defhandler button_clicked(_socket, _) do
+  def button_clicked(socket, sender, arg \\ "outside") do
+    # IO.inspect(socket)
+    set_prop(socket, this_commander(sender) <> " .spaceholder1", innerText: "changed")
+    poke(socket, text: "changed in shared commander, " <> arg, bgcolor: "#77dddd", color: "#990000")
   end
 
-  # def page_loaded(socket) do
-  #   DrabTestApp.IntegrationCase.add_page_loaded_indicator(socket)
-  #   DrabTestApp.IntegrationCase.add_pid(socket)
-  # end
+  defhandler peek_text(socket, sender, _) do
+    set_prop(socket, this(sender), innerText: peek(socket, :text) || "--- nil ---")
+  end
 end